crecion de servicio

+++ proyecto sistema operativos 2/src/main/java/com/umg/backend/repository/empleadoRepository.java
@@ -3,6 +3,6 @@ package com.umg.backend.repository;
 import com.umg.backend.entity.Empleado;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface empleadoRepository extends JpaRepository <Empleado, Integer> {
+public interface empleadoRepository extends JpaRepository <Empleado, Long> {
 
 }
